fix: resolve critical bugs and improve codebase quality
This commit addresses multiple critical, high, and medium severity issues
identified during codebase analysis, and adds comprehensive setup documentation.

CRITICAL FIXES:
- Fixed missing 'plan' variable in dashboard component (app/dashboard/page.tsx:31)
  Now properly destructures plan from usePlanLimits hook

- Eliminated N+1 query in response submission (app/api/forms/[id]/responses/route.ts)
  Now fetches form with user relation in single query using Prisma include

- Mitigated race condition in response creation
  Wrapped response creation and count increment in atomic transaction
  Added documentation about remaining race condition in limit checking

HIGH PRIORITY IMPROVEMENTS:
- Consolidated PLAN_LIMITS into shared constants file (lib/constants.ts)
  Eliminated duplicate definitions across multiple API routes
  Added type-safe helper functions: isValidPlan(), isUpgradeablePlan()
  Moved PLAN_PRICES to constants for consistency

- Improved type safety in payment routes
  Added proper plan validation using isUpgradeablePlan()
  Prevents invalid plan values from being stored in database

- Updated all imports to use shared constants
  Modified: app/api/forms/route.ts, app/api/forms/[id]/responses/route.ts,
  app/api/payments/create-order/route.ts, app/api/payments/verify/route.ts,
  lib/hooks/use-plan-limits.ts

DOCUMENTATION ENHANCEMENTS:
- Added Quick Start section for minimal local setup
  Developers can now run app with just PostgreSQL and Google OAuth

- Added comprehensive troubleshooting guide covering:
  * Database connection issues
  * Prisma migration errors
  * Authentication problems
  * Google Sheets integration issues
  * File upload errors
  * Common development issues

- Added development tips section with:
  * Prisma Studio usage
  * Type checking commands
  * Testing strategies
  * Plan limit testing SQL queries

- Clarified prerequisites and optional services
  Made it clear which features require which integrations

PERFORMANCE IMPROVEMENTS:
- Reduced database queries in response submission from 2 to 1
- Made response creation + count update atomic via transaction

FILES CHANGED:
- lib/constants.ts (NEW) - Shared plan limits and pricing configuration
- app/dashboard/page.tsx - Fixed missing plan variable
- app/api/forms/[id]/responses/route.ts - N+1 fix, race condition mitigation
- app/api/forms/route.ts - Use shared constants
- app/api/payments/create-order/route.ts - Use shared constants + validation
- app/api/payments/verify/route.ts - Added plan validation
- lib/hooks/use-plan-limits.ts - Use shared constants
- README.md - Comprehensive local setup and troubleshooting documentation

TESTING RECOMMENDATIONS:
- Test concurrent response submissions to verify transaction behavior
- Verify plan limit enforcement with different user plans
- Test minimal setup without optional integrations
- Follow troubleshooting guide to ensure accuracy
